Add files via upload
The price of a car depends on a lot of factors like the goodwill of the brand of the car, features of the car, horsepower and the mileage it gives and many more. 
Car price prediction is one of the major research areas in machine learning. 
I've made use of libraries such as pandas (to read in dataset files), numpy (to work with arrays used within the code), matplotlib & seaborn (for plotting), sklearn (for various ML models).
The decision tree regression algorithm to train a car price prediction model was used.
The data was split into training and test sets to train the model.  
I've also added the dataset used by me. 
The dataset was downloaded from Kaggle.com